pyproject.toml
[project]
name = "arco-2-0-pce"
version = "0.1.0"
description = ""
authors = [
    {name = "Paulo Cardoso",email = "cardosopaulohenrique91@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp (>=3.12.15,<4.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "duckdb (>=1.4.0,<2.0.0)",
    "google-api-python-client (>=2.183.0,<3.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "arco_2"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
